### 2021-04-17 #3179 @whedon
[PRE REVIEW]: The Babelizer: language interoperability for model coupling in the geosciences
**Submitting author:** @mcflugen (<a href="http://orcid.org/0000-0002-5864-6459">ERIC HUTTON</a>)
**Repository:** <a href="https://github.com/csdms/babelizer/" target ="_blank">https://github.com/csdms/babelizer/</a>
**Version:** v1.0.0
