### 2021-04-19 #3179 @kthyng
@mcflugen, @kyleniemeyer I'm not sure I have any relevant expertise for this. Geosciences, yes, but that looks like the application but not the guts of this project? @mcflugen who would you recommend to review this submission (from your network or [this list](https://bit.ly/joss-reviewers)? What type of skills are you looking for?
