+++ docs/source/babel_heatc.toml
@@ -24,3 +24,7 @@ package_author_email = "csdms@colorado.edu"
 package_license = "MIT"
 summary = "PyMT plugin for heat model"
 
+[ci]
+python_version = ["3.9"]
+os = ["linux", "mac", "windows"]
+
