### 2021-04-24 75b3667
Update docs for API changes (#47)
* Update long example for API changes

* Move headings; add acknowledgment

* Spell check

* Update changelog
