+++ docs/source/api/babelizer.rst
@@ -52,6 +52,7 @@ babelizer.wrap module
    :undoc-members:
    :show-inheritance:
 
+
 Module contents
 ---------------
 
