### 2021-05-03 d44994c
Revert "Format conf.py with black"
This reverts commit c14f914d55c91191d57da776bfdb6c00e22cb9be.
