### 2021-05-03 57e2475
Include docs/conf.py in list of files to format
