+++ setup.py
@@ -14,7 +14,7 @@ long_description = u"\n\n".join(
 
 setup(
     name="babelizer",
-    version="0.3.8.dev0",
+    version="0.3.8",
     description="Wrap bmi libraries with Python bindings",
     long_description=long_description,
     author="Eric Hutton",
