+++ CHANGES.rst
@@ -3,6 +3,12 @@ Changelog for babelizer
 =======================
 
 ******************
+0.3.9 (unreleased)
+******************
+
+- Nothing changed yet.
+
+
 0.3.8 (2021-05-03)
 ******************
 
