### 2021-06-01 #3179 @kthyng
@dbuscombe-usgs and @ngoldbaum: would you be interested in reviewing this submission for JOSS? We're asking for people to start their reviews in the next few weeks in hopes the whole iterative process would be finished in 4-6 weeks. You can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Thanks for your consideration!
