### 2021-06-04 #3179 @kthyng
@PythonCHB Are you available to review this JOSS submission? We're asking for people to start their reviews in the next few weeks in hopes the whole iterative process would be finished in 4-6 weeks. As a reminder, you can learn more about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Thanks for your consideration!
