### 2021-06-08 #3344 @whedon
[REVIEW]: The Babelizer: language interoperability for model coupling in the geosciences
**Submitting author:** @mcflugen (<a href="http://orcid.org/0000-0002-5864-6459">ERIC HUTTON</a>)
**Repository:** https://github.com/csdms/babelizer/
**Branch with paper.md** (empty if default branch): develop
**Version:** v0.3.9
