### 2021-06-22 csdms/babelizer#50 @lheagy
JOSS review: Readme suggestions 
These are a few suggestions for the README + intro page on the docs as a part of the JOSS review (openjournals/joss-reviews#3344). The README has a great, easy to follow set of instructions for installation! What I think could be improved though is the description of the package and who should be using it (e.g. to satisfy the "Statement of need" requirement in the JOSS documentation checklist). I appreciate it is a part of the larger CSDMS workbench, so a bit more context in the high level summary on where this project fits would be helpful for users to get a sense of if they are in the right place or if they should be looking at another repository in the CSDMS ecosystem. The summary in the paper is a good overview!
