### 2021-06-23 csdms/babelizer#51 @lheagy
JOSS review: suggestions on the example
For the JOSS review (openjournals/joss-reviews#3344), I am working through the example, and as I work through, I will list a few questions / suggestions here:
- rather than assuming that the `build` directory is in the home directory, it would be more convenient to always have relative paths (e.g. going back up / down from wherever you are currently)
-  for the testing step that uses the config file in the [Wrap the model with the babelizer](https://babelizer.readthedocs.io/en/latest/example.html#wrap-the-model-with-the-babelizer) stage (last step): I would suggest you walk the user through creating that file from the command line (e.g. `echo "1.5, 8.0, 6, 5" > config.txt`) possibly in addition to providing the link to the config file that you can download. It wasn't obvious that the text in the docs was hyperlinked, and it is a bit of messing around to get the config file into the right directory for such a simple file
- the directory structure gets a little involved, so it would be helpful to have a high-level summary / picture at the top to give an idea of where we are headed and the role of each directory. e.g. something like the image below is all I have in mind, and possibly with a couple of annotations of what each is
![image](https://user-images.githubusercontent.com/6361812/123016919-7b218100-d380-11eb-823a-1e64589c4e86.png)

but all ran smoothly and the instructions are clear! Nice example :)
