### 2021-06-23 csdms/babelizer#52 @lheagy
JOSS review: API documentation 
The [cli module](https://babelizer.readthedocs.io/en/latest/api/babelizer.html) is well documented, but the other modules are a bit sparse -- I suspect most users will be interacting through the cli, so this is the most important part, but even short descriptions associated with the rest or the public API would be nice to help with a picture of what can be accomplished with `babelizer`

linking with openjournals/joss-reviews#3344
