### 2021-06-23 csdms/babelizer#53 @lheagy
JOSS review: suggestion -- test coverage
Part of openjournals/joss-reviews#3344. This is just a suggestion and I don't think it is necessary for acceptance with JOSS, but it might be nice to have code coverage with the testing to provide an indication of how much of the codebase is touched by the tests. For example codecov integrates nicely with github actions: https://github.com/codecov/codecov-action.
