### 2021-06-23 #3344 @lheagy
Thanks for inviting me to review! This is a very cool project and I think an important contribution for enabling interoperability between codes and different sub-domains in the geosciences.

I linked a few issues with comments and suggestions. Please feel free to let me know if anything is unclear.

The only other suggestion that I have with respect to the documentation / paper is that it might be nice to include a bit of commentary on how this differes from python binding libraries (e.g. `f2py` or `pybind11`). These are common tools that come to mind when thinking about interoperability between languages, so I think discussing them directly would help build a better mental model of the role of `babelizer`.
