### 2021-08-10 csdms/babelizer#51 @mdpiper
Hi @lheagy! Thank you for your careful reading and evaluation of the example. I'll implement all your suggestions. PR forthcoming soon-ish.
