### 2021-08-10 f888dd3
Update example to use relative paths
