### 2021-08-10 csdms/babelizer#51 @mdpiper
@lheagy Please let me know if #54 addresses your suggestions above. The docs build for the example is up at https://babelizer.readthedocs.io/en/mdpiper-fix-51/example.html.
