### 2021-08-11 7ce27bb
Help autodoc find babelizer package
