+++ docs/source/conf.py
@@ -4,17 +4,17 @@
 # list see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
+import datetime
+import pkg_resources
+
 # -- Path setup --------------------------------------------------------------
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
-#
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
-import datetime
-import pkg_resources
+import os
+import sys
+sys.path.insert(0, os.path.abspath('../..'))
 
 
 # The master toctree document.
