### 2021-08-11 28dc83b
Use sphinx-click to autodoc babelizer.cli
It works wonderfully. See https://sphinx-click.readthedocs.io.
