### 2021-08-12 31cbe28
Edit docstrings in babelizer.cli module
I tried to make them look good at the CLI and in the docs.
