### 2021-08-12 2617ec7
Add docstrings to exception classes
