+++ babelizer/errors.py
@@ -1,4 +1,7 @@
 class BabelizeError(Exception):
+
+    """An exception that the babelizer can handle and show to the user."""
+
     def __init__(self, message):
         self._message = message
 
@@ -7,25 +10,25 @@ class BabelizeError(Exception):
 
 
 class ValidationError(BabelizeError):
-
+    """Raised if babelizer configuration data are incorrectly formatted."""
     pass
 
 
 class RenderError(BabelizeError):
-
+    """Raised if the babelizer encounters an error in creating output files."""
     pass
 
 
 class ScanError(BabelizeError):
-
+    """Raised if babelizer configuration file is incorrectly formatted."""
     pass
 
 
 class OutputDirExistsError(BabelizeError):
-
+    """An exception used when the directory for babelized output exists."""
     pass
 
 
 class SetupPyError(BabelizeError):
-
+    """Raised if the babelized package cannot be built through setup.py."""
     pass
