### 2021-08-12 64faba4
Ensure autodoc sees docstring in ctr
