+++ docs/source/api/babelizer.rst
@@ -27,6 +27,7 @@ babelizer.metadata module
 .. automodule:: babelizer.metadata
    :members:
    :undoc-members:
+   :special-members: __init__
    :show-inheritance:
 
 babelizer.render module
