### 2021-08-12 5770ce3
Add docstrings to metadata module
