### 2021-08-12 3c31062
Add docstrings for render and utils modules
