+++ babelizer/errors.py
@@ -11,24 +11,29 @@ class BabelizeError(Exception):
 
 class ValidationError(BabelizeError):
     """Raised if babelizer configuration data are incorrectly formatted."""
+
     pass
 
 
 class RenderError(BabelizeError):
     """Raised if the babelizer encounters an error in creating output files."""
+
     pass
 
 
 class ScanError(BabelizeError):
     """Raised if babelizer configuration file is incorrectly formatted."""
+
     pass
 
 
 class OutputDirExistsError(BabelizeError):
     """An exception used when the directory for babelized output exists."""
+
     pass
 
 
 class SetupPyError(BabelizeError):
     """Raised if the babelized package cannot be built through setup.py."""
+
     pass
