+++ docs/source/conf.py
@@ -14,7 +14,8 @@ import pkg_resources
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../..'))
+
+sys.path.insert(0, os.path.abspath("../.."))
 
 
 # The master toctree document.
