### 2021-08-12 csdms/babelizer#52 @mdpiper
Hi @lheagy! I have a candidate PR (#55) to fix this issue. The docs are [here](https://babelizer.readthedocs.io/en/mdpiper-fix-52/api/babelizer.html). Please let me know what you think!
