### 2021-08-13 be48f59
Remove non-language tests
