### 2021-08-13 e5c8c1e
Make new rule for language tests
