### 2021-08-16 2d15d37
Remove single-page API docs
