+++ docs/source/api/babelizer.rst
@@ -1,47 +1,0 @@
-babelizer package
-=================
-
-babelizer.cli module
---------------------
-
-.. automodule:: babelizer.cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. click:: babelizer.cli:babelize
-   :prog: babelize
-   :nested: full
-
-babelizer.errors module
------------------------
-
-.. automodule:: babelizer.errors
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-babelizer.metadata module
--------------------------
-
-.. automodule:: babelizer.metadata
-   :members:
-   :undoc-members:
-   :special-members: __init__
-   :show-inheritance:
-
-babelizer.render module
------------------------
-
-.. automodule:: babelizer.render
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-babelizer.utils module
-----------------------
-
-.. automodule:: babelizer.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
