+++ Makefile
@@ -74,9 +74,9 @@ coverage: ## check code coverage quickly with the default Python
 	$(BROWSER) htmlcov/index.html
 
 docs: ## generate Sphinx HTML documentation, including API docs
+	sphinx-apidoc -o docs/source/api babelizer --separate
 	rm -f docs/source/api/babelizer.rst
 	rm -f docs/source/api/modules.rst
-	sphinx-apidoc -o docs/source/api babelizer
 	$(MAKE) -C docs clean
 	$(MAKE) -C docs html
 	$(BROWSER) docs/build/html/index.html
