+++ docs/source/api/babelizer.wrap.rst
@@ -1,0 +1,10 @@
+babelizer.wrap module
+=====================
+
+.. deprecated:: 0.2
+   Use `babelizer.render.render` instead.
+
+.. automodule:: babelizer.wrap
+   :members:
+   :undoc-members:
+   :show-inheritance:
