### 2021-08-16 f0d138a
Tidying after building and reviewing docs
