+++ docs/source/api/babelizer.wrap.rst
@@ -2,7 +2,7 @@ babelizer.wrap module
 =====================
 
 .. deprecated:: 0.2
-   Use `babelizer.render.render` instead.
+   Use `babelizer.render` instead.
 
 .. automodule:: babelizer.wrap
    :members:
