### 2021-08-16 8785941
Complete API documentation (#55)
* Help autodoc find babelizer package

* Use sphinx-click to autodoc babelizer.cli

It works wonderfully. See https://sphinx-click.readthedocs.io.

* Remove unused Submodules and Module contents headings

* Edit docstrings in babelizer.cli module

I tried to make them look good at the CLI and in the docs.

* Add docstrings to exception classes

* Ensure autodoc sees docstring in ctr

* Add docstrings to metadata module

* The wrap module is no longer used

* Add docstrings for render and utils modules

* Make pretty with unforgiving black

* Be sure to blacken 
[…]
