+++ docs/source/api/babelizer.rst
@@ -1,62 +1,0 @@
-babelizer package
-=================
-
-Submodules
-----------
-
-babelizer.cli module
---------------------
-
-.. automodule:: babelizer.cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-babelizer.errors module
------------------------
-
-.. automodule:: babelizer.errors
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-babelizer.metadata module
--------------------------
-
-.. automodule:: babelizer.metadata
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-babelizer.render module
------------------------
-
-.. automodule:: babelizer.render
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-babelizer.utils module
-----------------------
-
-.. automodule:: babelizer.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-babelizer.wrap module
----------------------
-
-.. automodule:: babelizer.wrap
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: babelizer
-   :members:
-   :undoc-members:
-   :show-inheritance:
