+++ docs/source/api/babelizer.utils.rst
@@ -1,0 +1,7 @@
+babelizer.utils module
+======================
+
+.. automodule:: babelizer.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
