### 2021-08-16 ba0690e
Update and clarify example (#54)
* Add final directory structure image

* Tweak summary in sample babelizer input file

* Update example to use relative paths
