+++ docs/source/babel_heatc.toml
@@ -22,7 +22,7 @@ github_username = "pymt-lab"
 package_author = "csdms"
 package_author_email = "csdms@colorado.edu"
 package_license = "MIT"
-summary = "PyMT plugin for heat model"
+summary = "PyMT plugin for the C heat model"
 
 [ci]
 python_version = ["3.9"]
