### 2021-08-17 b7e2443
add "is the babelizer for me?" section
