+++ README.rst
@@ -16,12 +16,10 @@ The Babelizer: Wrap BMI libraries with Python bindings
 ======================================================
 
 
-The *babelizer* is a utility for wrapping libraries, from a variety of
-languages, that expose a `Basic Model Interface`_ (BMI) so that they can be
+The *babelizer* is a utility for wrapping a library that exposes a `Basic Model Interface`_ (BMI) so that it can be
 imported as a Python package.
 
-
-Supported languages:
+Supported languages include:
 
 *  C
 *  C++
@@ -36,58 +34,57 @@ The Babelizer is part of the CSDMS Workbench
 The *babelizer* is an element of the `CSDMS Workbench`_,
 an integrated system of software tools, technologies, and standards
 for building and coupling models. The Workbench provides two Python
-frameworks for model coupling, *pymt* and *landlab*, and so the *babelizer*
-was written to bring models written in other languages into
-these frameworks and, so, into Python. The *babelizer*, however,
-can be used outside of the CSDMS Workbench. So long as your model
+frameworks for model coupling, *pymt* and *landlab*.
+The *babelizer* was written to bring models written in other languages into
+these frameworks.
+However, as long as your model
 satisfies the requirements below, you can use the *babelizer*
 to bring your model into Python without having to use any of the
-other tools on the Workbench.
+other tools in the Workbench.
 
 
 ***************************
 Should I use the babelizer?
 ***************************
 
-We wrote the *babelizer* for model developers, who have written a model in languages other
-than Python, to use their models within a Python environment. To determine if the
-*babelizer* is right for you, you must first be aware of several requirements.
+To determine if the
+*babelizer* is right for you, first be aware of a few requirements.
 
-1. Your model must be written in one of the following languages: C, C++, Fortran, or Python
+1. Your model must be written in C, C++, Fortran, or Python
 2. Your model must provide a shared library
-3. Your model must expose a Basic Model Interface through this library
+3. Your model must expose a `Basic Model Interface`_ through this library
 
-The most difficult of the three requirements is the last--implementing a BMI. This will
-involve restructuring your code to implement a series of functions with prescribed names,
-arguments and return values for quering and controlling your model. We have created
-several resources that you can use to help you understand the BMI and to guide you
+The most difficult of the three requirements is the last--implementing a BMI. This
+involves adding a series of functions with prescribed names,
+arguments, and return values for querying and controlling your model. We have created
+several resources to help you understand the BMI and to guide you
 through the implementation process.
 
-BMI resources:
+BMI resources
+=============
 
 * The `Basic Model Interface`_ documentation provides an overview of the BMI as well
   as a detailed description of all of the BMI functions.
-* For each of the supported languages,
-* The following resources provide a BMI specification for each of the supported languages,
+* The following provide a BMI specification for each of the supported languages:
 
   * `C spec <https://github.com/csdms/bmi-c/>`_
   * `C++ spec <https://github.com/csdms/bmi-cxx/>`_
   * `Fortran spec <https://github.com/csdms/bmi-fortran/>`_
   * `Python spec <https://github.com/csdms/bmi-python/>`_
 
-* The following resources give examples of a BMI implementation for each of the supported languages,
+* The following give examples of a BMI implementation for each of the supported languages:
 
   * `C example <https://github.com/csdms/bmi-example-c/>`_
   * `C++ example <https://github.com/csdms/bmi-example-cxx/>`_
   * `Fortran example <https://github.com/csdms/bmi-example-fortran/>`_
   * `Python example <https://github.com/csdms/bmi-example-python/>`_
 
+Note
+====
 
-.. note::
-
-  There are lots of other good reasons to create a BMI for
[…]
