### 2021-09-09 04f7c1e
docs: improve the readme documentation (#57)
* add "is the babelizer for me?" section

* Minor edits

Co-authored-by: Mark Piper <mark.piper@colorado.edu>
