### 2021-09-09 feaa5d5
tests: update tests and generate coverage (#56)
* Generate coverage report

* Include coverage package

* Move language tests to a new directory

* Move language tests next to language examples

* Use a persistent tmpdir to test babelize update

* Remove non-language tests

* Add basic CLI tests

* Make new rule for language tests

* Generate coverage for tests

* Suppress builds of internal PRs
