+++ .github/workflows/build-test-ci.yml
@@ -4,6 +4,13 @@ on: [push, pull_request]
 
 jobs:
   build-and-test:
+    # We want to run on external PRs, but not on our own internal PRs as they'll be run
+    # by the push to the branch. Without this if check, checks are duplicated since
+    # internal PRs match both the push and pull_request events.
+    if:
+      github.event_name == 'push' || github.event.pull_request.head.repo.full_name !=
+      github.repository
+
     runs-on: ${{ matrix.os }}
 
     defaults:
@@ -74,6 +81,7 @@ jobs:
           make install
           popd
 
-      - name: Test language examples
+      - name: Test babelizer + language examples and generate coverage report
         run: |
-          pytest -vvv
+          coverage run --source babelizer --omit */babelizer/data/* -m pytest tests external/tests
+          coverage report -m
