+++ Makefile
@@ -59,7 +59,10 @@ pretty: ## reformat files to make them look pretty
 	black setup.py babelizer tests docs/source/conf.py --exclude=babelizer/data
 
 test: ## run tests quickly with the default Python
-	pytest --cov=babelizer
+	pytest tests --disable-warnings -vvv
+
+test-languages: ## run tests on babelizer languages
+	pytest external/tests --disable-warnings -vvv
 
 benchmark: ## run benchmarks only
 	pytest --benchmark-only --benchmark-autosave
@@ -68,7 +71,7 @@ test-all: ## run tests on every Python version with tox
 	tox
 
 coverage: ## check code coverage quickly with the default Python
-	coverage run --source babelizer -m pytest
+	coverage run --source babelizer --omit */babelizer/data/* -m pytest tests external/tests
 	coverage report -m
 	coverage html
 	$(BROWSER) htmlcov/index.html
