### 2021-09-09 csdms/babelizer#50 @mdpiper
Hi @lheagy -- @mcflugen attempted to address this issue in #57. (It may be easier to view the resulting [README.rst](https://github.com/csdms/babelizer/blob/develop/README.rst) in the *develop* branch.) Please let us know if these changes satisfy the concerns raised in this issue.
