### 2021-09-14 97c0d4b
Write tests for babelize generate
