### 2021-09-14 59b11ef
Norm metadata: set default branch if github_branch not present
