+++ babelizer/render.py
@@ -93,7 +93,7 @@ def render_plugin_repo(template, context=None, output_dir=".", clobber=False):
     if not path.is_dir():
         raise RenderError("error creating {0}".format(path))
 
-    git.Repo.init(path)
+    git.Repo.init(path, initial_branch=context["info"]["github_branch"])
 
     return path
 
