### 2021-09-16 e4b09fe
Update lint and pretty rules to include external/tests
