### 2021-09-16 51df6d5
Norm default max_instances value
