### 2021-09-16 5a1ecd8
Use max_instances build attribute in Fortran interoperability code
