### 2021-09-16 7152795
Create tests for new max_instances build parameter
