+++ external/tests/test_fortran/user-instances.toml
@@ -1,0 +1,30 @@
+[library]
+[library.HeatBMI]
+language = "fortran"
+library = "bmiheatf"
+header = ""
+entry_point = "bmi_heat"
+
+[build]
+undef_macros = []
+define_macros = []
+libraries = []
+library_dirs = []
+include_dirs = []
+extra_compile_args = []
+max_instances = 32
+
+[package]
+name = "pymt_heatf"
+requirements = [""]
+
+[info]
+github_username = "pymt-lab"
+package_author = "csdms"
+package_author_email = "csdms@colorado.edu"
+package_license = "MIT"
+summary = "PyMT plugin for heatf model"
+
+[ci]
+python_version = ["3.9"]
+os = ["linux", "mac", "windows"]
