### 2021-10-14 750cdfc
parametrize github_branch checks
