### 2021-10-14 9fdaac7
added test dependencies for babelize init working
