+++ external/tests/test_python.py
@@ -18,6 +18,24 @@ def sessiondir(tmpdir_factory):
     return sdir
 
 
+@pytest.mark.dependency()
+def test_babelize_init_works(tmpdir, datadir):
+    runner = CliRunner(mix_stderr=False)
+    with tmpdir.as_cwd():
+        result = runner.invoke(babelize, ["init", str(datadir / "babel.toml")])
+
+        assert result.stdout.strip() == str(tmpdir / "pymt_heatpy")
+        assert result.exit_code == 0
+        assert pathlib.Path("pymt_heatpy").exists()
+        assert (pathlib.Path("pymt_heatpy") / "babel.toml").is_file()
+
+        repo = git.Repo("pymt_heatpy")
+
+        assert not repo.bare
+        assert repo.active_branch.name == "main"
+
+
+@pytest.mark.dependency(depends=["test_babelize_init_works"])
 @pytest.mark.parametrize("branch", ["main", "master", "coffee", None])
 def test_babelize_init_python_with_branch(tmpdir, datadir, branch):
     runner = CliRunner(mix_stderr=False)
@@ -46,8 +64,14 @@ def test_babelize_init_python_with_branch(tmpdir, datadir, branch):
         assert repo.active_branch.name == branch
 
 
-def test_babelize_build_python_example(sessiondir, datadir):
-    with sessiondir.as_cwd():
+@pytest.mark.dependency(depends=["test_babelize_init_works"])
+def test_babelize_build_python_example(tmpdir, datadir):
+    runner = CliRunner(mix_stderr=False)
+
+    with tmpdir.as_cwd():
+        runner.invoke(babelize, ["init", str(datadir / "babel.toml")])
+
+    with tmpdir.as_cwd():
         try:
             result = subprocess.run(
                 ["pip", "install", "-e", "."],
@@ -88,12 +112,16 @@ def test_babelize_build_python_example(sessiondir, datadir):
         assert result.returncode == 0
 
 
-def test_babelize_update_python(sessiondir, datadir):
-    runner = CliRunner()
+@pytest.mark.dependency(depends=["test_babelize_init_works"])
+def test_babelize_update_python(tmpdir, datadir):
+    runner = CliRunner(mix_stderr=False)
 
-    with sessiondir.as_cwd():
+    with tmpdir.as_cwd():
+        runner.invoke(babelize, ["init", str(datadir / "babel.toml")])
+
+    with tmpdir.as_cwd():
         result = runner.invoke(babelize, ["--cd", "pymt_heatpy", "update"])
 
         assert result.exit_code == 0
         assert pathlib.Path("pymt_heatpy").exists()
-        assert "re-rendering" in result.output
+        assert "re-rendering" in result.stderr
