### 2021-10-21 bdf540e
Removed unused babelizer config file
