+++ external/tests/test_python/user-branch.toml
@@ -1,30 +1,0 @@
-[library]
-[library.HeatPy]
-language = "python"
-library = "heat"
-header = ""
-entry_point = "BmiHeat"
-
-[build]
-undef_macros = []
-define_macros = []
-libraries = []
-library_dirs = []
-include_dirs = []
-extra_compile_args = []
-
-[package]
-name = "pymt_heatpy"
-requirements = []
-
-[info]
-github_username = "csdms"
-github_branch = "coffee"
-package_author = "CSDMS"
-package_author_email = "csdms@colorado.edu"
-package_license = "MIT"
-summary = "PyMT component for the heatpy model"
-
-[ci]
-python_version = ["3.9"]
-os = ["linux", "mac", "windows"]
