### 2021-10-21 292db13
Remove unused babelizer config file
