### 2021-10-21 1fae6cb
Parameterize max_instances in test and mark init test as dependency
