+++ external/tests/test_fortran.py
@@ -4,19 +4,48 @@ import pathlib
 import shutil
 import subprocess
 import tomlkit as toml
+import pytest
 
 from click.testing import CliRunner
 
 from babelizer.cli import babelize
 
 
-def test_babelize_init_fortran_default_instances(tmpdir, datadir):
-    runner = CliRunner()
+@pytest.mark.dependency()
+def test_babelize_init_works(tmpdir, datadir):
+    runner = CliRunner(mix_stderr=False)
+    with tmpdir.as_cwd():
+        result = runner.invoke(babelize, ["init", str(datadir / "babel.toml")])
+
+        assert result.stdout.strip() == str(tmpdir / "pymt_heatf")
+        assert result.exit_code == 0
+        assert pathlib.Path("pymt_heatf").exists()
+        assert (pathlib.Path("pymt_heatf") / "babel.toml").is_file()
+
+        repo = git.Repo("pymt_heatf")
+
+        assert not repo.bare
+
+
+@pytest.mark.dependency(depends=["test_babelize_init_works"])
+@pytest.mark.parametrize("max_instances", [8, 32, None])
+def test_babelize_init_fortran_with_user_instances(tmpdir, datadir, max_instances):
+    runner = CliRunner(mix_stderr=False)
+
+    babel_toml = toml.load(datadir / "babel.toml")
+    babel_toml["build"]["max_instances"] = max_instances
+
+    if max_instances is None:
+        babel_toml["build"].pop("max_instances", None)
+        max_instances = 8
 
     with tmpdir.as_cwd():
-        shutil.copy(datadir / "babel.toml", ".")
+        with open("babel.toml", "w") as fp:
+            toml.dump(babel_toml, fp)
+
         result = runner.invoke(babelize, ["init", "babel.toml"])
 
+        assert result.stdout.strip() == str(tmpdir / "pymt_heatf")
         assert result.exit_code == 0
         assert pathlib.Path("pymt_heatf").exists()
         assert (pathlib.Path("pymt_heatf") / "babel.toml").is_file()
@@ -25,34 +54,17 @@ def test_babelize_init_fortran_default_instances(tmpdir, datadir):
             meta = toml.parse(fp.read())
 
         assert "max_instances" in meta["build"]
-        assert meta["build"]["max_instances"] == 8
+        assert meta["build"]["max_instances"] == max_instances
 
 
-def test_babelize_init_fortran_user_instances(tmpdir, datadir):
-    runner = CliRunner()
-
-    with tmpdir.as_cwd():
-        shutil.copy(datadir / "user-instances.toml", ".")
-        result = runner.invoke(babelize, ["init", "user-instances.toml"])
-
-        assert result.exit_code == 0
-        assert pathlib.Path("pymt_heatf").exists()
-        assert (pathlib.Path("pymt_heatf") / "babel.toml").is_file()
-
-        with open("pymt_heatf/babel.toml", "r") as fp:
-            meta = toml.parse(fp.read())
-
-        assert "max_instances" in meta["build"]
-        assert meta["build"]["max_instances"] == 32
-
-
+@pytest.mark.dependency(depends=["test_babelize_init_works"])
 def test_babelize_build_fortran_example(tmpdir, datadir):
-    runner = CliRunner()
+    runner = CliRunner(mix_stderr=False)
 
     with tmpdir.as_cwd():
-        shutil.copy(datadir / "babel.toml", ".")
-        result = runner.invoke(babelize, ["init", "babel.toml"])
+        runner.invoke(babelize, ["init", str(datadir / "babel.toml")])
 
+    with tmpdir.as_cwd():
         try:
             result = subprocess.run(
[…]
