### 2021-11-08 csdms/babelizer#62 @cheginit
[JOSS Review] Supported Python versions
Although in `README` the supported Python version is stated as `>=3.8` but in the `setup.py` or `setup.cfg` and the `conda-forge` recipe, the required fields are not specified. This could potentially lead to issues when installing the package with other Python packages that may not support Python 3.8+ versions.

For `setup.py` or `setup.cfg` the field is `python_requires='>3.8'` and for `conda-forge` recipe, i.e., `meta.yml` file, `- python >= 3.8` should be specified in both `host` and `run` sections.

This issue is related to openjournals/joss-reviews#3344
