### 2021-11-08 csdms/babelizer#62 @cheginit
I just noticed that `bmi-tester` does not support Python 3.10 so you might want to note that Python 3.10 is not supported. At least for development purposes.
