### 2021-11-08 #3344 @cheginit
I just started the review process.

Overall, I find this submission a much-needed and useful contribution to the ESM community. The paper provides a detailed overview of the statement of the need and their approach for addressing their goals. As it's mentioned in the paper the biggest hurdle for taking advantage of this package is the need for models to follow the BMI framework. Nevertheless, having a package such as `babelizer` might encourage developers to consider the BMI approach in their development pipeline. So I think this package is a step toward a right direction in the ESM community. During my initial review I found some issue. I opened two issues. I will continue updating it.

Comments on the paper:
[ ] L28: There seems to be extra commas in "new, unsolved, problem". Maybe you meant "new and unsolved problems" or "new unsolved problems".
[ ] The paper doesn't discuss if there's been any similar effort, whether in Python or any other language.
