### 2022-01-20 7b136df
add some initial text about similar previous works
