### 2022-01-20 b3308ce
add Python 3.10 to ci tests
