+++ .github/workflows/build-test-ci.yml
@@ -19,8 +19,10 @@ jobs:
 
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
-        python-version: ["3.8", "3.9", "3.10"]
+        # os: [ubuntu-latest, macos-latest]
+        # python-version: ["3.8", "3.9", "3.10"]
+        os: [macos-latest]
+        python-version: ["3.9"]
 
     steps:
       - uses: actions/checkout@v2
