+++ .github/workflows/build-test-ci.yml
@@ -20,9 +20,8 @@ jobs:
     strategy:
       matrix:
         # os: [ubuntu-latest, macos-latest]
-        # python-version: ["3.8", "3.9", "3.10"]
+        python-version: ["3.8", "3.9", "3.10"]
         os: [macos-latest]
-        python-version: ["3.9"]
 
     steps:
       - uses: actions/checkout@v2
