+++ .github/workflows/build-test-ci.yml
@@ -20,8 +20,8 @@ jobs:
     strategy:
       matrix:
         # os: [ubuntu-latest, macos-latest]
-        python-version: ["3.8", "3.9", "3.10"]
-        os: [macos-latest]
+        python-version: ["3.9"]
+        os: [ubuntu-latest]
 
     steps:
       - uses: actions/checkout@v2
