### 2022-01-20 8b1aadb
add Python 3.10 back to tests
