+++ .github/workflows/build-test-ci.yml
@@ -20,7 +20,7 @@ jobs:
     strategy:
       matrix:
         # os: [ubuntu-latest, macos-latest]
-        python-version: ["3.9"]
+        python-version: ["3.9", "3.10"]
         os: [ubuntu-latest]
 
     steps:
