### 2022-01-20 0f78c48
require python >= 3.9; test on linux and mac
