### 2022-01-20 5c75a39
tests: add Python 3.10 to ci tests, drop Python 3.8 (#65)
* add Python 3.10 to ci tests

* temporarily limit builds for debugging

* temporarily only test on mac

* temporarily build on only linux

* add Python 3.10 back to tests

* add python 3.8 to builds

* require python >= 3.9; test on linux and mac

* add a requirements file for development

* remove unused requirements
