+++ requirements.txt
@@ -1,9 +1,7 @@
 black
 click
 gitpython
-mock
 pyyaml
 tomlkit
 isort>=5
-jinja2
 cookiecutter
