+++ paper/paper.md
@@ -115,7 +115,7 @@ but, rather, a code generator that produces glue code to provide cross-language
 interoperability. The generated code is then passed to a traditional compiler
 to build libraries.
 
-Whereas the `babelizer` uses Python as a hub language that connects it’s
+Whereas the `babelizer` uses Python as a hub language that connects its
 supported languages, babel itself acts as the hub, generating spokes to each
 supported language so that languages talk directly to one another. That is,
 babel is capable of generating bridges from each of its supported languages
@@ -124,10 +124,13 @@ interfaces while the *babelizer* is only able to wrap libraries that expose a BM
 
 ### Grpc4BMI
 
-Several other groups have adopted an alternative solution to the language
-interoperability problem that treats models as web services that expose a BMI.
-In this framework, models are built on a server or within a software container
-(e.g. Docker) and interact with clients through network ports.
+Another alternative solution to the language interoperability problem is to treat
+models as web services that expose a BMI (e.g. Grpc4BMI as described by [@hut:2021]).
+In such a framework, models are built on separate servers or within their own
+software container (e.g. Docker) and interact with clients through network ports.
+Isolating models within environments eliminates the potential of dependency
+conflicts between models and requires them to only be built within on specific
+environment or operating system.
 
 
 ![The distribution of programming languages used in the models in the CSDMS model repository. (Data from https://csdms.colorado.edu/wiki/CSDMS_models_by_numbers.)\label{fig:languages}](language_fig.png)
