### 2022-02-15 611e2e7
Text edits; moved Figure 1 closer to reference
Hopefully moving the figure won't upset the page formatting.
