### 2022-02-15 0f63e18
Move text before page break forced by Fig 2
