### 2022-02-15 4f57dd1
Fix citation formatting
