+++ paper/paper.bib
@@ -100,7 +100,7 @@
 
 @article{hut:2021,
   title={The eWaterCycle platform for Open and FAIR Hydrological collaboration},
-  author={Hut, Rolf and Drost, Niels and van de Giesen, Nick and van Werkhoven, Ben and Abdollahi, Banafsheh and Aerts, Jerom and Albers, Thomas and Alidoost, Fakhereh and Andela, Bouwe and Camphuijsen, Jaro and others},
+  author={Hut, Rolf and Drost, Niels and van de Giesen, Nick and van Werkhoven, Ben and Abdollahi, Banafsheh and Aerts, Jerom and Albers, Thomas and Alidoost, Fakhereh and Andela, Bouwe and Camphuijsen, Jaro and Yifat Dzigan and Ronald van Haren and Eric Hutton and Peter Kalverla and Maarten van Meersbergen and Gijs van den Oord and Inti Pelupessy and Stef Smeets and Stefan Verhoeven and Martine de Vos and Berend Weel},
   journal={Geoscientific Model Development Discussions},
   pages={1--31},
   year={2021},
