### 2022-02-16 edc7d7e
Add text to paper describing similar work (#64)
* Update docs for API changes (#47)

* Update long example for API changes

* Move headings; add acknowledgment

* Spell check

* Update changelog

* Fix minor formatting issues (#49)

* Swallow empty line before sample import statement

* Include docs/conf.py in list of files to format

* Format with black

* Update changelog

* Add conda-forge badge

* Preparing release 0.3.8

* Back to development: 0.3.9

* Complete API documentation (#55)

* Help autodoc find babelizer package

* Use sphinx-click to autodoc babelizer.cli

It works wonderfully
[…]
