+++ paper/paper.bib
@@ -86,3 +86,23 @@
   doi = {10.31223/X51615},
   url = {https://doi.org/10.31223/X51615}
 }
+
+@article{epperly:2012,
+  title={High-performance language interoperability for scientific computing through Babel},
+  author={Epperly, Thomas GW and Kumfert, Gary and Dahlgren, Tamara and Ebner, Dietmar and Leek, Jim and Prantl, Adrian and Kohn, Scott},
+  journal={The International Journal of High Performance Computing Applications},
+  volume={26},
+  number={3},
+  pages={260--274},
+  year={2012},
+  publisher={Sage Publications Sage UK: London, England}
+}
+
+@article{hut:2021,
+  title={The eWaterCycle platform for Open and FAIR Hydrological collaboration},
+  author={Hut, Rolf and Drost, Niels and van de Giesen, Nick and van Werkhoven, Ben and Abdollahi, Banafsheh and Aerts, Jerom and Albers, Thomas and Alidoost, Fakhereh and Andela, Bouwe and Camphuijsen, Jaro and Yifat Dzigan and Ronald van Haren and Eric Hutton and Peter Kalverla and Maarten van Meersbergen and Gijs van den Oord and Inti Pelupessy and Stef Smeets and Stefan Verhoeven and Martine de Vos and Berend Weel},
+  journal={Geoscientific Model Development Discussions},
+  pages={1--31},
+  year={2021},
+  publisher={Copernicus GmbH}
+}
