+++ paper/paper.md
@@ -32,19 +32,18 @@ bibliography: paper.bib
 
 # Summary
 
-The `babelizer` is a Python utility that generates code
+The *babelizer* is a Python utility that generates code
 to import libraries from other languages into Python. Target libraries
 must expose a Basic Model Interface (BMI) [@hutton:2020b; @peckham:2013] and be written in
-C, C++, or Fortran, although the `babelizer` is extendable, so
-other languages can be added in the future. The `babelizer` provides a
+C, C++, or Fortran, although the *babelizer* is extendable, so
+other languages can be added in the future. The *babelizer* provides a
 streamlined mechanism for bringing scientific models into a common language
 where they can communicate with one another as components of an integrated model.
 
 
 # Statement of need
 
-With an integrated multicomponent approach to modeling, scientific
-modelers--not just software developers--connect components
+With an integrated multicomponent approach to modeling, scientists--not just software developers--connect components
 to form integrated models, where plug-and-play
 components can easily be added or removed [@tucker:2021; @david:2013; @gregersen:2007; @collins:2005].
 This is in contrast to older methods, where a single modeling group would construct
@@ -69,18 +68,18 @@ greater risk that community-developed models will become incompatible
 with one another. With hundreds of scientists developing models in
 isolation, there is a greater likelihood models will be written with
 idiosyncratic designs, incompatible grids, incompatible time steps,
-and even in different programming languages. The Earth-system modeling
+and in different programming languages. The Earth-surface modeling
 community has developed tools to help solve some of these problems.
 For example, the Basic Model Interface
 standardizes model interactions. The Earth System Modeling Framework (ESMF) [@collins:2005]
 grid mappers are able to map quantities from one grid to another.
-The Python Modeling Toolkit `pymt` [@hutton:2020a] performs time
+The Python Modeling Toolkit *pymt* [@hutton:2020a] performs time
 interpolation, grid mapping, and unit coversion.
 In this paper, we present a solution to the language incompatibility problem.
 
 ## Overcoming the language incompatibility problem
 
-To get an idea of the range of programming languages used in Earth-system
+To get an idea of the range of programming languages used in Earth-surface
 modeling, we can look to the Community Surface Dynamics Modeling System (CSDMS)
 model repository. As of June 2020, the repository holds over 370 open source
 models and tools submitted by the community. These contributions span a range of languages, with Python, C, and Fortran
@@ -92,9 +91,14 @@ connects to other languages.
 We chose Python as
 the hub because of its popularity in the scientific community,
 its extensive collection of third-party libraries (including model
-coupling frameworks such as the `pymt`), and its existing ability to
+coupling frameworks such as the *pymt*), and its existing ability to
 communicate with other programming languages.
-We have built the `babelizer` to generate the spokes that connect Python to other languages.
+We have built the *babelizer* to generate the spokes that connect Python to other languages.
+
+
+![The distribution of programming languages used in the models in the CSDMS model repository. (Data from https://csdms.colorado.edu/wiki/CSDMS_models_by_numbers.)\label{fig:languages}](language_fig.png)
+
+
 Using the CSDMS model repository as a
 guide, if we build translators for the open
 source languages C, C++, Fortran, and Python, we will cover 80 percent of
@@ -103,36 +107,58 @@ A drawback of using
 Python is that it can be relatively slow compared to compiled
 languages like C and Fortran; however, the models being wrapped
 are compiled and run in their native language, which is where
-the bulk of the computation takes place, with the `babelizer` providing
+the bulk of the computation takes place, with the *babelizer* providing
 only a thin wrapper layer.
 
[…]
