### 2022-02-16 f92e171
ci: update black GitHub action (#67)
* update black GitHub action

* update style with new version of black
