+++ .github/workflows/black.yml
@@ -18,4 +18,5 @@ jobs:
       - uses: actions/setup-python@v2
       - uses: psf/black@stable
         with:
-          args: ". --check"
+          options: "--check --verbose --diff"
+          src: "setup.py babelizer tests"
