### 2022-02-16 ed118b7
set --omit default for coverage
