+++ pyproject.toml
@@ -18,6 +18,9 @@ doctest_optionflags = [
     "ALLOW_UNICODE"
 ]
 
+[tool.coverage.run]
+omit = "*/babelizer/data/*"
+
 [tool.isort]
 multi_line_output = 3
 include_trailing_comma = true
