### 2022-02-16 2751972
change omit option to a list of strings
