### 2022-02-17 9ac5176
use setup.cfg instead of pyproject for coverage config
