### 2022-02-17 435f3dd
docs: add coverage badge to readme (#66)
* add coverage badge to readme

* add coverage packages to testing requirements

* build docs with Python 3.9

* use pytest-cov, upload to coveralls

* set --omit default for coverage

* change omit option to a list of strings

* install coverage[toml]

* use setup.cfg instead of pyproject for coverage config
