+++ .github/workflows/docs.yml
@@ -38,7 +38,7 @@ jobs:
           - uses: conda-incubator/setup-miniconda@v2
             with:
               auto-update-conda: true
-              python-version: 3.8
+              python-version: 3.9
               channels: conda-forge
               channel-priority: true
 
