+++ setup.cfg
@@ -8,3 +8,7 @@ max-line-length = 88
 
 [zest.releaser]
 tag-format = v{version}
+
+[coverage:run]
+omit = */babelizer/data/*
+relative_files = True
