+++ external/requirements.txt
@@ -1,0 +1,10 @@
+bmi-tester>=0.5.4
+bmi-c
+bmi-cxx
+bmi-fortran
+bmipy
+cmake
+c-compiler
+cxx-compiler
+fortran-compiler
+pkg-config
