### 2022-02-17 35c956a
Use Tucker et al. GMD article instead of preprint (#70)
