+++ paper/paper.bib
@@ -78,13 +78,16 @@
   url = {https://doi.org/10.1002/essoar.10504855.1}
 }
 
-@article{tucker:2021,
-  title={Numerical modeling of Earth's dynamic surface: a community approach},
-  author={Tucker, Gregory E and Hutton, Eric and Piper, Mark and Campforts, Benjamin and Gan, Tian and Barnhart, Katherine R and Kettner, Albert and Overeem, Irina and Peckham, Scott D and McCready, Lynn and Syvitski, Jaia},
-  year={2021},
-  journal={EarthArXiv},
-  doi = {10.31223/X51615},
-  url = {https://doi.org/10.31223/X51615}
+@article{tucker:2022,
+  author = {Tucker, G. E. and Hutton, E. W. H. and Piper, M. D. and Campforts, B. and Gan, T. and Barnhart, K. R. and Kettner, A. J. and Overeem, I. and Peckham, S. D. and McCready, L. and Syvitski, J.},
+  title = {CSDMS: a community platform for numerical modeling of Earth surface processes},
+  journal = {Geoscientific Model Development},
+  volume = {15},
+  year = {2022},
+  number = {4},
+  pages = {1413--1439},
+  url = {https://gmd.copernicus.org/articles/15/1413/2022/},
+  doi = {10.5194/gmd-15-1413-2022}
 }
 
 @article{epperly:2012,
