### 2022-02-17 c2bfa9d
dev: reorganize the testing requirements (#68)
* remove conda requirements for testing languages

* add requirements file for testing languages

* update readme to include pip and conda install

* update ci to test languages and use new requirements file
