### 2022-02-24 #3344 @mcflugen
@cheginit We added two new sections, "*The Common Component Architecture*" and "*Grpc4BMI*" that discuss similar efforts to solve the language interoperability problem that the *babelizer* tries to address. We think this addresses your second suggestion.

I've also removed the extra commas on line 28. I just did this now, sorry I missed it earlier.

Were there other suggestions you had?
