### 2022-03-01 #3344 @kthyng
@mcflugen I'm happy to say that this is ready to move forward toward publishing! I'm very sorry that the JOSS process is taken so long for this submission. It's not common but clearly does happen sometimes.

Your paper looks good. Next steps are:
1. Confirm what version you'd like associated with this JOSS paper and maybe make a new release to that effect.
2. Upload an archive of your software to a place like Zenodo and report back here the DOI. Also please update the title and author list so that it exactly matches your JOSS paper, for consistency.
