### 2022-03-04 197335b
docs: added JOSS paper (#44)
* add paper.md for JOSS submission

* update to almost a first draft

* Add bibfile; cite Peckham, Hutton

I want to check that these render correctly.

* Editing pass (#38)

* Add DOI for Peckham ref

* Add ESMF and pymt refs

* Editing pass 1

Still need to make figures.

* Minor edits

* add script to make language pie graph

* add target to build joss paper

* docs: add references and figures (#42)

* Edit text

* Add references for component modeling

* Add sample babel.toml file for PRMSSurface

* Fix typo

* Add Syvitski to author list

* Change citation sty
[…]
