+++ paper/paper.bib
@@ -1,0 +1,111 @@
+@article{collins:2005,
+  title={Design and implementation of components in the Earth System Modeling Framework},
+  author={Collins, Nancy and Theurich, Gerhard and Deluca, Cecelia and Suarez, Max and Trayanov, Atanas and Balaji, Venkatramani and Li, Peggy and Yang, Weiyu and Hill, Chris and Da Silva, Arlindo},
+  journal={The International Journal of High Performance Computing Applications},
+  volume={19},
+  number={3},
+  pages={341--350},
+  year={2005},
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+}
+
+@article{david:2013,
+  title = {A software engineering perspective on environmental modeling framework design: The Object Modeling System},
+  journal = {Environmental Modelling & Software},
+  volume = {39},
+  pages = {201-213},
+  year = {2013},
+  doi = {10.1016/j.envsoft.2012.03.006},
+  url = {https://doi.org/10.1016/j.envsoft.2012.03.006},
+  author = {O. David and J.C. Ascough and W. Lloyd and T.R. Green and K.W. Rojas and G.H. Leavesley and L.R. Ahuja}
+}
+
+@article{gregersen:2007,
+  title={OpenMI: Open modelling interface},
+  author={Gregersen, JB and Gijsbers, PJA and Westen, SJP},
+  journal={Journal of Hydroinformatics},
+  volume={9},
+  number={3},
+  pages={175--191},
+  year={2007},
+  doi = {10.2166/hydro.2007.023},
+  url = {https://doi.org/10.2166/hydro.2007.023}
+}
+
+@software{hutton:2020a,
+  author       = {Eric W. H. Hutton and Mark D. Piper},
+  title        = {The Python Modeling Toolkit},
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v1.0.0},
+  doi          = {10.5281/zenodo.3644240},
+  url          = {https://doi.org/10.5281/zenodo.3644240}
+}
+
+@article{hutton:2020b,
+  doi = {10.21105/joss.02317},
+  url = {https://doi.org/10.21105/joss.02317},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {51},
+  pages = {2317},
+  author = {Eric W. H. Hutton and Mark D. Piper and Gregory E. Tucker},
+  title = {The Basic Model Interface 2.0: A standard interface for coupling numerical models in the geosciences},
+  journal = {Journal of Open Source Software}
+}
+
+@article{peckham:2013,
+  title={A component-based approach to integrated modeling in the
+         geosciences: The design of CSDMS},
+  author={Peckham, Scott D and Hutton, Eric WH and Norris, Boyana},
+  journal={Computers \& Geosciences},
+  volume={53},
+  pages={3--12},
+  year={2013},
+  publisher={Elsevier},
+  doi = {10.1016/j.cageo.2012.04.002},
+  url = {https://doi.org/10.1016/j.cageo.2012.04.002}
+}
+
+@article{piper:2020,
+  title={Coupling Hydrologic Models with Data Services in an Interoperable Modeling Framework},
+  author={Piper, Mark and McDonald, Richard and Hutton, Eric and Markstrom, Steven and Parker, Norton and Tucker, Gregory},
+  journal={Earth and Space Science Open Archive ESSOAr},
+  year={2020},
+  publisher={American Geophysical Union},
+  doi = {10.1002/essoar.10504855.1},
+  url = {https://doi.org/10.1002/essoar.10504855.1}
+}
+
+@article{tucker:2022,
[…]
