+++ CHANGES.rst
@@ -1,101 +1,146 @@
-=======================
-Changelog for babelizer
-=======================
+Changelog
+=========
 
-******************
-0.3.9 (unreleased)
-******************
+.. towncrier release notes start
 
-- Complete API documentation (#55)
-
-
-******************
 0.3.8 (2021-05-03)
-******************
+------------------
 
-- Update docs for API changes (#47)
-- Fix minor formatting issues in rendered output (#49)
+Bugfixes
+""""""""
+
+- Fix minor formatting issues in rendered output. (`#49
+  <https://github.com/csdms/babelizer/issues/49>`_)
 
 
-******************
+Improved Documentation
+""""""""""""""""""""""
+
+- Update docs for API changes. (`#47
+  <https://github.com/csdms/babelizer/issues/47>`_)
+
+
 0.3.7 (2021-03-19)
-******************
+------------------
 
-- Render GitHub Actions for continuous integration of the generated
-  projects (#41)
+Features
+""""""""
 
-- Use GitHub Actions for continuous integration (#39)
+- Render GitHub Actions for continuous integration of the generated projects.
+  (`#41 <https://github.com/csdms/babelizer/issues/41>`_)
 
 
-******************
+Misc
+""""
+
+- Use GitHub Actions for continuous integration. (`#39
+  <https://github.com/csdms/babelizer/issues/39>`_)
+
+
 0.3.6 (2021-01-13)
-******************
+------------------
 
-- Fixed issue with metadata install path for Python components (#36)
+Bugfixes
+""""""""
 
-- Minor edits to README
+- Fixed issue with metadata install path for Python components. (`#36
+  <https://github.com/csdms/babelizer/issues/36>`_)
 
 
-******************
 0.3.5 (2020-12-16)
-******************
+------------------
 
-- Removed use of the toml package, instead use tomlkit (#34)
+Misc
+""""
+
+- Removed use of the toml package, instead use tomlkit. (`#34
+  <https://github.com/csdms/babelizer/issues/34>`_)
 
 
-******************
 0.3.4 (2020-12-16)
-******************
+------------------
 
-- Minor edits to README and CLI help strings (#29)
+Bugfixes
+""""""""
 
-- Added text on the CSDMS Workbench to the README and docs (#30)
-
-- Clarified text in docs example (#31)
-
-- Fixed a templating problem that caused Python BMIs to fail to build.
-  Included the Python BMI example in the test suite. (#33)
+- Fixed a templating problem that caused Python BMIs to fail to build. (`#33
+  <https://github.com/csdms/babelizer/issues/33>`_)
 
 
-******************
+Improved Documentation
+""""""""""""""""""""""
+
+- Minor edits to README and CLI help strings. (`#29
+  <https://github.com/csdms/babelizer/issues/29>`_)
+- Added text on the CSDMS Workbench to the README and docs. (`#30
+  <https://github.com/csdms/babelizer/issues/30>`_)
+- Clarified text in docs example. (`#31
+  <https://github.com/csdms/babelizer/issues/31>`_)
+
+
+Misc
+""""
+
+- Included the Python BMI example in the test suite.. (`#33
+  <https://github.com/csdms/babelizer/issues/33>`_)
+
+
 0.3.3 (2020-10-31)
-******************
+------------------
+
+Features
+""""""""
 
 - Added missing methods, primarily for unstructured grids, to C and C++
-  implementation (#28)
-
-- Fixed a rendering error that caused import lines to run together
-  when wrapping multiple components (#28)
-
-- Removed Python 3.6 builds (#28)
+  implementation. (`#28 <https://github.com/csdms/babelizer/issues/28>`_)
 
 
-******************
+Bugfixes
+""""""""
+
[…]
