+++ README.rst
@@ -15,7 +15,6 @@
     :target: https://coveralls.io/github/csdms/babelizer?branch=develop
 
 
-======================================================
 The Babelizer: Wrap BMI libraries with Python bindings
 ======================================================
 
@@ -31,9 +30,8 @@ Supported languages include:
 *  Python
 
 
-********************************************
 The Babelizer is part of the CSDMS Workbench
-********************************************
+--------------------------------------------
 
 The *babelizer* is an element of the `CSDMS Workbench`_,
 an integrated system of software tools, technologies, and standards
@@ -47,9 +45,8 @@ to bring your model into Python without having to use any of the
 other tools in the Workbench.
 
 
-***************************
 Should I use the babelizer?
-***************************
+---------------------------
 
 To determine if the
 *babelizer* is right for you, first be aware of a few requirements.
@@ -65,7 +62,7 @@ several resources to help you understand the BMI and to guide you
 through the implementation process.
 
 BMI resources
-=============
+^^^^^^^^^^^^^
 
 * The `Basic Model Interface`_ documentation provides an overview of the BMI as well
   as a detailed description of all of the BMI functions.
@@ -84,16 +81,15 @@ BMI resources
   * `Python example <https://github.com/csdms/bmi-example-python/>`_
 
 Note
-====
+^^^^
 
 There are lots of other good reasons to create a BMI for
 your model--not just so you can bring it into Python with the *babelizer*!
 Read all about them in the `Basic Model Interface`_ documentation.
 
 
-************
 Requirements
-************
+------------
 
 The *babelizer* requires Python >=3.9.
 
@@ -110,9 +106,9 @@ additional dependencies for running the *babelizer*'s tests to make sure
 that things are working as they should. These dependencies are listed
 in *requirements-testing.txt*.
 
-************
+
 Installation
-************
+------------
 
 To install the *babelizer*, first create a new environment.
 Although this isn't strictly necessary, it
@@ -125,7 +121,7 @@ base Python installation. This can be done with *conda*:
     $ conda activate babelizer
 
 Stable Release
-==============
+^^^^^^^^^^^^^^
 
 The *babelizer* and its dependencies are best installed with *conda*:
 
@@ -134,7 +130,7 @@ The *babelizer* and its dependencies are best installed with *conda*:
     $ conda install babelizer -c conda-forge
 
 From Source
-===========
+^^^^^^^^^^^
 
 After downloading the the *babelizer* source code, run the following from
 *babelizer*'s top-level directory (the one that contains *setup.py*) to
@@ -150,9 +146,9 @@ or using *conda*:
 
     $ conda install --file=requirements.txt -c conda-forge
 
-**********
+
 Input file
-**********
+----------
 
 The *babelizer* requires a single *toml*-formatted input file that describes
 the library to wrap. This file is typically named *babel.toml*.
@@ -198,19 +194,19 @@ For example, the above *babel.toml* was generated with:
   $ babelize generate > babel.toml
 
 Library section
-===============
+^^^^^^^^^^^^^^^
 
 The *library* section specifies information about the library being babelized.
 
 Name
-----
+""""
 
 The name of the babelized class.
 This will be a Python class,
 so it should follow Python naming conventions such as camel-case typing.
 
 Language
---------
+""""""""
 
 The programming language of the library (possible values are "c", "c++",
 "fortran", and "python").
@@ -221,20 +217,20 @@ The programming language of the library (possible values are "c", "c++",
   language = "c"
 
 Library
--------
+"""""""
 
 The name of the BMI library to wrap.
 This is the text passed to the linker through the `-l` option;
[…]
