### 2022-03-04 afc809c
update the manifest to exclude paper folder
