+++ pyproject.toml
@@ -34,3 +34,13 @@ exclude = '''
   | external
 )
 '''
+
+[tool.towncrier]
+directory = "news"
+package = "babelizer"
+filename = "CHANGES.rst"
+single_file = true
+underlines = "-^\""
+issue_format = "`#{issue} <https://github.com/csdms/babelizer/issues/{issue}>`_"
+title_format = "{version} ({project_date})"
+wrap = true
