+++ CHANGES.rst
@@ -3,6 +3,12 @@ Changelog
 
 .. towncrier release notes start
 
+0.3.10 (unreleased)
+-------------------
+
+- Nothing changed yet.
+
+
 0.3.9 (2022-03-04)
 ------------------
 
