### 2022-03-04 #3344 @mcflugen
@kthyng No worries about the delay, it was an unusual submission that made it difficult to find reviewers and part of the delay was definitely on our side.

Anyway, the version we would like associated with this JOSS paper is *v0.3.9*. I've created a release, and uploaded it to Zenodo.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6329737.svg)](https://doi.org/10.5281/zenodo.6329737)
