### 2022-03-17 0b2a2b3
Add missing DOIs (#71)
