+++ paper/paper.bib
@@ -6,6 +6,8 @@
   number={3},
   pages={341--350},
   year={2005},
+  doi = {10.1177/1094342005056120},
+  url = {https://doi.org/10.1177/1094342005056120},
   publisher={Sage Publications Sage CA: Thousand Oaks, CA}
 }
 
@@ -98,6 +100,8 @@
   number={3},
   pages={260--274},
   year={2012},
+  doi = {10.1177/1094342011414036},
+  url = {https://doi.org/10.1177/1094342011414036},
   publisher={Sage Publications Sage UK: London, England}
 }
 
