### 2022-03-17 3f112f6
add python 3.8 to the ci tests
