+++ .github/workflows/build-test-ci.yml
@@ -19,7 +19,8 @@ jobs:
 
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        # os: [ubuntu-latest, macos-latest]
+        os: [macos-latest]
         python-version: ["3.8", "3.9", "3.10"]
 
     steps:
