+++ .github/workflows/build-test-ci.yml
@@ -19,8 +19,7 @@ jobs:
 
     strategy:
       matrix:
-        # os: [ubuntu-latest, macos-latest]
-        os: [macos-latest]
+        os: [ubuntu-latest, macos-latest]
         python-version: ["3.8", "3.9", "3.10"]
 
     steps:
@@ -41,6 +40,10 @@ jobs:
           pip install coverage[toml]
           mamba install --file=requirements.txt --file=requirements-testing.txt --file=external/requirements.txt
           conda list
+
+      - name: Install build system
+        run: |
+          mamba install cython numpy setuptools wheel
 
       - name: Install babelizer
         run: |
