### 2022-03-17 2e9c5dc
install c example with --no-build-isolation
