### 2022-03-17 92f1c12
tests: add python 3.8 to the ci tests (#72)
* add python 3.8 to the ci tests

* require python>=3.8

* disable linux builds

* install the build system with conda

* install c example with --no-build-isolation

* install c, fortran examples with --no-build-isolation

* add news fragment

[skip ci]

* isolate builds except with linux and python <= 3.8

* convert python minor version to integer
